refactor: split secrets scanner modules
